third commit
changed lot of things..

First of all, a structure where the user can use role-based get post and delete endpoints was required.

Created the basis of auth with jwt token and issued authorization-based jwt token and made transactions with that token.

We had two roles named user and admin and we stored the data of which user these roles belong to with 2 tables (user, user-role) on the database side. After the user posts the login method with username and password, the jwt token is generated for the role and can perform transactions with that token.