fourth commit
Basically, a modular structure was switched to a modular structure,

we had previously opened 2 tables on the database and on the user_role table we were keeping a data about which authorization the users received, instead of that, with a modular design, we now keep the authorizations we have created on the table and we mark the users in the users table according to their roles using a foreign key.
During the authorization check, we take the user's role id and now we check which role it corresponds to on the table whose name is roles and we make the controller operations by checking which role it corresponds to.